Update tutorial-tuple-spaces.md
Class `Tuple` in Java is used internally in jSpace and is not part of the public interface of put/get/etc. methods. The previous tutorial was suggesting to do things like 

```
space.put(new Tuple("hello"));
```

which would actually insert a tuple whose first field is a tuple!

This has now been fixed in the tutorial.